增加dry-run参数用于调试脚本

+++ build.sh
@@ -7,30 +7,33 @@ ROOT=`realpath -s $SCRIPT_DIR/..`
 FLASH_TOOL=$ROOT/tools/linux/Linux_Upgrade_Tool/Linux_Upgrade_Tool/upgrade_tool
 PACK_FIRMWARE_PATH=${ROOT}/tools/linux/Linux_Pack_Firmware/rockdev
 
-# $1: single camera or not
-# $2: enable adb or not
-# $3: full image or not
-# $4: clean all before build
+# arguments
+FOR_586=0
+ENABLE_ADB=0
+FULL_IMG=0
+CLEAN=0
+DRY_RUN=0
+
 build() {
 	# info print
 	echo "printing basic info..."
 	echo ">>> buildroot dir: $ROOT"
-	if [ $1 == 0 ]; then
+	if [ $FOR_586 == 0 ]; then
 		echo ">>> build for: imx334/imx678"
 	else
 		echo ">>> build for: imx586"
 	fi
-	if [ $2 == 0 ]; then
+	if [ $ENABLE_ADB == 0 ]; then
 		echo ">>> adb enabled: false"
 	else
 		echo ">>> adb disabled: true"
 	fi
-	if [ $3 == 0 ]; then
+	if [ $FULL_IMG == 0 ]; then
 		echo ">>> build type: incremental upgrade image"
 	else
 		echo ">>> build type: full image"
 	fi
-	if [ $4 == 0 ]; then
+	if [ $CLEAN == 0 ]; then
 		echo ">>> clean all before build: false"
 	else
 		echo ">>> clean all before build: true"
@@ -38,7 +41,7 @@ build() {
 	echo ""
 
 	# basic check
-	if check_prerequisite $1 $2; then
+	if check_prerequisite; then
 		echo "check prerequisite ok"
 	else
 		echo "check prerequisite failed, abort building"
@@ -46,7 +49,7 @@ build() {
 	fi
 
 	# clean or not
-	if [ $4 == 1 ]; then
+	if [ $CLEAN == 1 ]; then
 		echo ""
 		echo "perform a clean build, clean all..."
 		cd $ROOT
@@ -63,23 +66,25 @@ build() {
 	# build rkscript
 	echo ""
 	echo "rebuilding rkscript for adb..."
-	rm -f $ROOT/buildroot/output/rockchip_rk3588/target/etc/init.d/S50usbdevice
-	cd $ROOT/buildroot
-	make rkscript-rebuild
+	build_rkscript
 
 	# build firmware
 	echo ""
 	echo "building firmware..."
-	cd $ROOT
-	./build.sh
+	if [ $DRY_RUN == 0 ]; then
+		cd $ROOT
+		./build.sh
+	fi
 
 	# generate image
 	echo ""
 	echo "generate firmware image..."
-	if [ $3 == 0 ]; then
-		gen_app_firmware
-	else
-		gen_system_firmware
+	if [ $DRY_RUN == 0 ]; then
+		if [ $FULL_IMG == 0 ]; then
+			gen_app_firmware
+		else
+			gen_system_firmware
+		fi
 	fi
 
 	# end
@@ -87,8 +92,14 @@ build() {
 	echo "all done!!"
 }
 
-# $1: single camera or not
-# $2: enable adb or not
+build_rkscript() {
+	if [ $DRY_RUN == 0 ]; then
+		rm -f $ROOT/buildroot/output/rockchip_rk3588/target/etc/init.d/S50usbdevice
+		cd $ROOT/buildroot
+		make rkscript-rebuild
+	fi
+}
+
 check_prerequisite() {
 	echo "check prerequisite..."
 
@@ -135,66 +146,86 @@ check_prerequisite() {
 		echo "checking imx334.c...already existed"
 	else
 		echo "checking imx334.c....copied"
-		cp $SCRIPT_DIR/files/imx334.c $ROOT/kernel/drivers/media/i2c
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/imx334.c $ROOT/kernel/drivers/media/i2c
+		fi
 	fi
 	if [ -f $ROOT/kernel/drivers/media/i2c/imx586.c ]; then
 		echo "checking imx586.c...already existed"
 	else
 		echo "checking imx586.c....copied"
-		cp $SCRIPT_DIR/files/imx586.c $ROOT/kernel/drivers/media/i2c
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/imx586.c $ROOT/kernel/drivers/media/i2c
+		fi
 	fi
 	if [ -f $ROOT/kernel/drivers/media/i2c/imx678.c ]; then
 		echo "checking imx678.c...already existed"
 	else
 		echo "checking imx678.c....copied"
-		cp $SCRIPT_DIR/files/imx678.c $ROOT/kernel/drivers/media/i2c
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/imx678.c $ROOT/kernel/drivers/media/i2c
+		fi
 	fi
 
 	# check configs
-	if [ $1 == 0 ]; then
-		cp -f $SCRIPT_DIR/files/for_334_678/rk3588_weewa.config $ROOT/kernel/arch/arm64/configs
-	else
-		cp -f $SCRIPT_DIR/files/for_586/rk3588_weewa.config $ROOT/kernel/arch/arm64/configs
+	if [ $DRY_RUN == 0 ]; then
+		if [ $FOR_586 == 0 ]; then
+			cp -f $SCRIPT_DIR/files/for_334_678/rk3588_weewa.config $ROOT/kernel/arch/arm64/configs
+		else
+			cp -f $SCRIPT_DIR/files/for_586/rk3588_weewa.config $ROOT/kernel/arch/arm64/configs
+		fi
 	fi
 	echo "checking rk3588 weewa config...copied"
 
 	# check dts
-	if [ $1 == 0 ]; then
-		cp -f $SCRIPT_DIR/files/for_334_678/rk3588-weewa-v10-linux.dts $ROOT/kernel/arch/arm64/boot/dts/rockchip
-	else
-		cp -f $SCRIPT_DIR/files/for_586/rk3588-weewa-v10-linux.dts $ROOT/kernel/arch/arm64/boot/dts/rockchip
+	if [ $DRY_RUN == 0 ]; then
+		if [ $FOR_586 == 0 ]; then
+			cp -f $SCRIPT_DIR/files/for_334_678/rk3588-weewa-v10-linux.dts $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		else
+			cp -f $SCRIPT_DIR/files/for_586/rk3588-weewa-v10-linux.dts $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		fi
 	fi
 	echo "checking rk3588-weewa-v10-linux.dts...copied"
 	if [ -f $ROOT/kernel/arch/arm64/boot/dts/rockchip/rk3588-weewa-v10.dtsi ]; then
 		echo "checking rk3588-weewa-v10.dtsi...already existed"
 	else
-		cp $SCRIPT_DIR/files/rk3588-weewa-v10.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/rk3588-weewa-v10.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		fi
 		echo "checking rk3588-weewa-v10.dtsi...copied"
 	fi
 	if [ -f $ROOT/kernel/arch/arm64/boot/dts/rockchip/rk3588-weewa-lp4-v10.dtsi ]; then
 		echo "checking rk3588-weewa-lp4-v10.dtsi...already existed"
 	else
-		cp $SCRIPT_DIR/files/rk3588-weewa-lp4-v10.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/rk3588-weewa-lp4-v10.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		fi
 		echo "checking rk3588-weewa-lp4-v10.dtsi...copied"
 	fi
 	if [ -f $ROOT/kernel/arch/arm64/boot/dts/rockchip/rk3588-weewa-cam-v10.dtsi ]; then
 		echo "checking rk3588-weewa-cam-v10.dtsi...already existed"
 	else
-		cp $SCRIPT_DIR/files/rk3588-weewa-cam-v10.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/rk3588-weewa-cam-v10.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		fi
 		echo "checking rk3588-weewa-cam-v10.dtsi...copied"
 	fi
 	if [ -f $ROOT/kernel/arch/arm64/boot/dts/rockchip/rk3588-weewa-cam-imx586-one.dtsi ]; then
 		echo "checking rk3588-weewa-cam-imx586-one.dtsi...already existed"
 	else
-		cp $SCRIPT_DIR/files/rk3588-weewa-cam-imx586-one.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		if [ $DRY_RUN == 0 ]; then
+			cp $SCRIPT_DIR/files/rk3588-weewa-cam-imx586-one.dtsi $ROOT/kernel/arch/arm64/boot/dts/rockchip
+		fi
 		echo "checking rk3588-weewa-cam-imx586-one.dtsi...copied"
 	fi
 
 	# check adb enabled or not
-	if [ $2 == 0 ]; then
-		cp -f $SCRIPT_DIR/files/disable_adb/rkscript.mk $ROOT/buildroot/package/rockchip/rkscript
-	else
-		cp -f $SCRIPT_DIR/files/enable_adb/rkscript.mk $ROOT/buildroot/package/rockchip/rkscript
+	if [ $DRY_RUN == 0 ]; then
+		if [ $ENABLE_ADB == 0 ]; then
+			cp -f $SCRIPT_DIR/files/disable_adb/rkscript.mk $ROOT/buildroot/package/rockchip/rkscript
+		else
+			cp -f $SCRIPT_DIR/files/enable_adb/rkscript.mk $ROOT/buildroot/package/rockchip/rkscript
+		fi
 	fi
 	echo "checking rkscript.mk...copied"
 
@@ -202,54 +233,57 @@ check_prerequisite() {
 }
 
 build_weewa_app() {
-	cd $ROOT
-	echo -n "46" | source envsetup.sh 
-	cd buildroot
-	set -x
-	echo -e "\033[32m[0] make weewa_idle-dirclean \033[0m"
-	make weewa_idle-dirclean
-	echo -e "\033[32m[1] weewa_idle-rebuild \033[0m"
-	make weewa_idle-rebuild
-	echo -e "\033[32m[2] weewa_daemon-dirclean \033[0m"
-	make weewa_daemon-dirclean
-	echo -e "\033[32m[3] make weewa_daemon-rebuild \033[0m"
-	make weewa_daemon-rebuild
-	echo -e "\033[32m[4] make weewa_daemon-dirclean \033[0m"
-	make rkipc-dirclean
-	echo -e "\033[32m[5] make rkipc-rebuild \033[0m"
-	make rkipc-rebuild
-	echo -e "\033[32m[6] build_weewa_app success \033[0m"
-	set +x
+	if [ $DRY_RUN == 0 ]; then
+		cd $ROOT
+		echo -n "46" | source envsetup.sh 
+		cd buildroot
+		set -x
+		echo -e "\033[32m[0] make weewa_idle-dirclean \033[0m"
+		make weewa_idle-dirclean
+		echo -e "\033[32m[1] weewa_idle-rebuild \033[0m"
+		make weewa_idle-rebuild
+		echo -e "\033[32m[2] weewa_daemon-dirclean \033[0m"
+		make weewa_daemon-dirclean
+		echo -e "\033[32m[3] make weewa_daemon-rebuild \033[0m"
+		make weewa_daemon-rebuild
+		echo -e "\033[32m[4] make weewa_daemon-dirclean \033[0m"
+		make rkipc-dirclean
+		echo -e "\033[32m[5] make rkipc-rebuild \033[0m"
+		make rkipc-rebuild
+		echo -e "\033[32m[6] build_weewa_app success \033[0m"
+		set +x
+	fi
 }
 
 gen_app_firmware() {
-	cd $PACK_FIRMWARE_PATH
-	cp $SCRIPT_DIR/files/package-file-app rk3588-package-file
-	./mkupdate.sh
-	datetime=$(date +"%Y%m%d")
+	if [ $DRY_RUN == 0 ]; then
+		cd $PACK_FIRMWARE_PATH
+		cp $SCRIPT_DIR/files/package-file-app rk3588-package-file
+		./mkupdate.sh
+		datetime=$(date +"%Y%m%d")
 
-	if [ ! -d $ROOT/firmware ];then
-	  mkdir $ROOT/firmware
+		if [ ! -d $ROOT/firmware ];then
+		  mkdir $ROOT/firmware
+		fi
+		mv update.img $ROOT/firmware/update-app-${datetime}.img
 	fi
-	mv update.img $ROOT/firmware/update-app-${datetime}.img
 }
 
 gen_system_firmware() {
-	cd $PACK_FIRMWARE_PATH
-	cp $SCRIPT_DIR/files/package-file-system rk3588-package-file
-	./mkupdate.sh
-	datetime=$(date +"%Y%m%d")
+	if [ $DRY_RUN == 0 ]; then
+		cd $PACK_FIRMWARE_PATH
+		cp $SCRIPT_DIR/files/package-file-system rk3588-package-file
+		./mkupdate.sh
+		datetime=$(date +"%Y%m%d")
 
-	if [ ! -d $ROOT/firmware ];then
-	  mkdir $ROOT/firmware
+		if [ ! -d $ROOT/firmware ];then
+		  mkdir $ROOT/firmware
+		fi
+		mv update.img $ROOT/firmware/update-system-${datetime}.img
 	fi
-	mv update.img $ROOT/firmware/update-system-${datetime}.img
 }
 
-FOR_586=0
-ENABLE_ADB=0
-FULL_IMG=0
-CLEAN=0
+# parse arguments
 for i in "$@"; do
 	case $i in
 		--586)
@@ -264,9 +298,12 @@ for i in "$@"; do
 	 	--clean)
 	 		CLEAN=1
 	 		;;
+	 	--dry-run)
+	 		DRY_RUN=1
+	 		;;
 	 	*)
 	 		;;
 	 esac
 done
 
-build $FOR_586 $ENABLE_ADB $FULL_IMG $CLEAN
+build
